Move builtin operators to dali/pipeline.
Move `ExternalSource`, `MakeContiguous` and `Copy` to dali/pipeline.
Move `DummyOp` to dali/test.
Add forced loading of libdali_operators by explicitly calling a function `InitOperatorsLib`.
