+++ dali/operators/operators.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2018, NVIDIA CORPORATION. All rights reserved.
+// Copyright (c) 2019, NVIDIA CORPORATION. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,19 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "dali/operators/util/dummy_op.h"
-
-#include <cstdlib>
+#include "dali/core/api_helper.h"
+#include "dali/operators/operators.h"
 
 namespace dali {
-
-DALI_REGISTER_OPERATOR(DummyOp, DummyOp<CPUBackend>, CPU);
-
-DALI_SCHEMA(DummyOp)
-  .DocStr("Dummy operator for testing")
-  .OutputFn([](const OpSpec &spec) { return spec.GetArgument<int>("num_outputs"); })
-  .NumInput(0, 10)
-  .AddOptionalArg("num_outputs",
-      R"code(Number of outputs.)code", 2);
-
+DLL_PUBLIC void InitOperatorsLib() {}
 }  // namespace dali
