+++ dali/operators/operators.h
@@ -1,0 +1,26 @@
+// Copyright (c) 2019, NVIDIA CORPORATION. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef DALI_OPERATORS_OPERATORS_H_
+#define DALI_OPERATORS_OPERATORS_H_
+
+namespace dali {
+/**
+ * @brief The function to reference, when one needs to make sure DALI operators shared
+ *        object is actually linked against.
+ */
+DLL_PUBLIC void InitOperatorsLib();
+}  // namespace dali
+
+#endif  // DALI_OPERATORS_OPERATORS_H_
