+++ dali/pipeline/operator/builtin/external_source.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "dali/operators/util/external_source.h"
+#include "dali/pipeline/operator/builtin/external_source.h"
 
 namespace dali {
