+++ dali/pipeline/operator/builtin/make_contiguous.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DALI_OPERATORS_UTIL_MAKE_CONTIGUOUS_H_
-#define DALI_OPERATORS_UTIL_MAKE_CONTIGUOUS_H_
+#ifndef DALI_PIPELINE_OPERATOR_BUILTIN_MAKE_CONTIGUOUS_H_
+#define DALI_PIPELINE_OPERATOR_BUILTIN_MAKE_CONTIGUOUS_H_
 
 #include <algorithm>
 #include <vector>
@@ -135,4 +135,4 @@ class MakeContiguous : public Operator<MixedBackend> {
 
 }  // namespace dali
 
-#endif  // DALI_OPERATORS_UTIL_MAKE_CONTIGUOUS_H_
+#endif  // DALI_PIPELINE_OPERATOR_BUILTIN_MAKE_CONTIGUOUS_H_
