+++ dali/pipeline/operator/op_schema.cc
@@ -25,6 +25,31 @@ std::map<string, OpSchema>& SchemaRegistry::registry() {
   return schema_map;
 }
 
+OpSchema& SchemaRegistry::RegisterSchema(const std::string &name) {
+  auto &schema_map = registry();
+  DALI_ENFORCE(schema_map.count(name) == 0, "OpSchema already "
+      "registered for operator '" + name + "'. DALI_SCHEMA(op) "
+      "should only be called once per op.");
+
+  // Insert the op schema and return a reference to it
+  schema_map.emplace(std::make_pair(name, OpSchema(name)));
+  return schema_map.at(name);
+}
+
+const OpSchema& SchemaRegistry::GetSchema(const std::string &name) {
+  auto &schema_map = registry();
+  auto it = schema_map.find(name);
+  DALI_ENFORCE(it != schema_map.end(), "Schema for operator '" +
+      name + "' not registered");
+  return it->second;
+}
+
+const OpSchema* SchemaRegistry::TryGetSchema(const std::string &name) {
+  auto &schema_map = registry();
+  auto it = schema_map.find(name);
+  return it != schema_map.end() ? &it->second : nullptr;
+}
+
 int OpSchema::CalculateOutputs(const OpSpec &spec) const {
   if (!output_fn_) {
     return num_output_;
