+++ dali/pipeline/pipeline_test.cc
@@ -22,7 +22,7 @@
 #include "dali/pipeline/data/buffer.h"
 #include "dali/pipeline/data/tensor.h"
 #include "dali/pipeline/operator/operator.h"
-#include "dali/operators/util/copy.h"
+#include "dali/pipeline/operator/builtin/copy.h"
 #include "dali/test/dali_test.h"
 #include "dali/test/dali_test_decoder.h"
 #include "dali/util/image.h"
