+++ dali/test/dummy_op.cu
@@ -12,16 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "dali/operators/util/make_contiguous.h"
+#include "dali/test/dummy_op.h"
 
 namespace dali {
 
-DALI_REGISTER_OPERATOR(MakeContiguous, MakeContiguous, Mixed);
-
-DALI_SCHEMA(MakeContiguous)
-  .DocStr(R"code(Move input batch to a contiguous representation, more suitable for execution on the GPU)code")
-  .NumInput(1)
-  .NumOutput(1)
-  .MakeInternal();
+DALI_REGISTER_OPERATOR(DummyOp, DummyOp<GPUBackend>, GPU);
 
 }  // namespace dali
